requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
sentence-transformers
openai
faiss-cpu
fpdf
fbprophet
herepy
pygame
sounddevice
numpy
scipy
speechrecognition
scikit-learn
gtts
pdfplumber
pandas
plotly.express
matplotlib
seaborn
kaleido 



